Description: install/rules fixes. Reviewed by: dcid Bug:

+++ src/config/localfile-config.c
@@ -28,14 +28,14 @@ int Read_Localfile(XML_NODE node, void *d1, void *d2)
     char *xml_localfile_location = "location";
     char *xml_localfile_logformat = "log_format";
 
-    logreader *log;
+    logreader *logf;
 
-    log = (logreader *)d1;
+    logf = (logreader *)d1;
     
-    log[MAX_READ_FILE].file = NULL;
-    log[MAX_READ_FILE].logformat = NULL;
+    logf[MAX_READ_FILE].file = NULL;
+    logf[MAX_READ_FILE].logformat = NULL;
     
-    while(log[pl].file != NULL)
+    while(logf[pl].file != NULL)
         pl++;
 
     if(pl >= MAX_READ_FILE)
@@ -44,9 +44,9 @@ int Read_Localfile(XML_NODE node, void *d1, void *d2)
         return(OS_INVALID);     
     }
     
-    log[pl].file = NULL;
-    log[pl].logformat = NULL;
-    log[pl].fp = NULL;
+    logf[pl].file = NULL;
+    logf[pl].logformat = NULL;
+    logf[pl].fp = NULL;
     
     /* Searching for entries related to files */
     i = 0;
@@ -64,29 +64,29 @@ int Read_Localfile(XML_NODE node, void *d1, void *d2)
         }
         else if(strcmp(node[i]->element,xml_localfile_location) == 0)
         {
-            os_strdup(node[i]->content, log[pl].file);                
+            os_strdup(node[i]->content, logf[pl].file);                
         }
 
         else if(strcasecmp(node[i]->element,xml_localfile_logformat) == 0)
         {
-            os_strdup(node[i]->content, log[pl].logformat);
+            os_strdup(node[i]->content, logf[pl].logformat);
 
-            if(strcmp(log[pl].logformat, "syslog") == 0)
+            if(strcmp(logf[pl].logformat, "syslog") == 0)
             {
             }
-            else if(strcmp(log[pl].logformat, "snort-full") == 0)
+            else if(strcmp(logf[pl].logformat, "snort-full") == 0)
             {
             }
-            else if(strcmp(log[pl].logformat, "snort-fast") == 0)
+            else if(strcmp(logf[pl].logformat, "snort-fast") == 0)
             {
             }
-            else if(strcmp(log[pl].logformat, "apache") == 0)
+            else if(strcmp(logf[pl].logformat, "apache") == 0)
             {
             }
-            else if(strcmp(log[pl].logformat, "squid") == 0)
+            else if(strcmp(logf[pl].logformat, "squid") == 0)
             {
             }
-            else if(strcmp(log[pl].logformat, EVENTLOG) == 0)
+            else if(strcmp(logf[pl].logformat, EVENTLOG) == 0)
             {
             }
             else
@@ -106,28 +106,28 @@ int Read_Localfile(XML_NODE node, void *d1, void *d2)
     }
 
     /* Missing log format */
-    if(!log[pl].logformat)
+    if(!logf[pl].logformat)
     {
         merror(MISS_LOG_FORMAT, ARGV0);
         return(OS_INVALID);
     }
 
     /* Missing file */
-    if(!log[pl].file)
+    if(!logf[pl].file)
     {
         merror(MISS_FILE, ARGV0);
         return(OS_INVALID);
     }
     
     /* Verifying a valid event log config */
-    if(strcmp(log[pl].logformat, EVENTLOG) == 0)
+    if(strcmp(logf[pl].logformat, EVENTLOG) == 0)
     {
-        if((strcmp(log[pl].file, "Application") != 0)&&
-           (strcmp(log[pl].file, "System") != 0)&&
-           (strcmp(log[pl].file, "Security") != 0))
+        if((strcmp(logf[pl].file, "Application") != 0)&&
+           (strcmp(logf[pl].file, "System") != 0)&&
+           (strcmp(logf[pl].file, "Security") != 0))
          {
              /* Invalid event log */
-             merror(INV_EVTLOG, ARGV0, log[pl].file);
+             merror(INV_EVTLOG, ARGV0, logf[pl].file);
              return(OS_INVALID);
          }
     }
